Remove do_shortcode from ajax handler, and add to unlocked template.
The parser class's get_shortcode_content function retrieves an array of content blocks, so it can't be fed to the do_shortcode function.

We loop through the content array, and apply the unlocked template to each content block, so I applied the do_shortcode function in template so control of output could be customized.
The content blocks weren't having the paragraphs added automatically via ajax like they do on page reload, so I explicitly applied the wpautop function to the content in the template.